==> 01_ETL_games.ipynb <==
# PROYECTO INTEGRADOR: STEAM
## Dataset - output_steam_games.json

## 1) Ingesta de datos

Se descomprimen cada uno de los archivos previo a importarlos con pandas.

#Importamos las librerías que utilizaremos a lo largo de nuestro proyecto
import pandas as pd
from pandas import json_normalize
from textblob import TextBlob
import json
import ast 


Importamos la data del archivo output_steam_games.json.

#Importamos en un dataframe la información sobre los juegos de Steam
with open("Data/output_steam_games.json",'r') as f:
    steam_games=[json.loads(line) for line in f]
df_steam_games=pd.DataFrame(steam_games)


#Observamos el dataframe
df_steam_games.head()


#Observamos las dimensiones de nuestro dataframe
df_steam_games.shape

## 2) Tratamiento de los datos

### 2.1) Dataframe: df_steam_games

Podemos observar que en el Dataframe df_steam_games los primeros 88310 registros, corresponden con los 88310 registros que se encuentran en el dataframe df_users_items, por lo tanto podemos proceder a eliminarlos.

#Verificamos los primeros 88310 datos y comparamos con df_items
df_steam_games.iloc[0:10]

#Ultimo dato a eliminar del dataframe
df_steam_games.iloc[88309]

#Tomamos el que debería ser el primer dato de nuestro dataframe
df_steam_games.iloc[88310]

#Elimino los primeros 88310 registros.
df_steam_games = df_steam_games.drop(index=range(88310))

#El primer dato de mi dataframe es el dato 88310 mostrado anteriormente. 
df_steam_games.iloc[0] 

#Verificamos las nuevas dimensiones de nuestro dataframe
df_steam_games.shape

#Reseteamos los indices
df_steam_games=df_steam_games.reset_index(drop=True)

#Sumamos la cantidad de valores nulos que hay en cada una de las columnas del df_steam_games 
df_steam_games.isnull().sum()

Vemos que las últimas columnas del dataframe (user_id, steam_id, items, items_count) están vacías, por lo que vamos a proceder a eliminarlas. 

#Eliminamos las columnas sin información
df_steam_games=df_steam_games.drop(columns=['user_id','steam_id', 'items', 'items_count'], axis=1)


#Verificamos las nuevas dimensiones del dataframe
df_steam_games.shape

Vamos a eliminar algunas de las columnas del dataframe que no usaremos a lo largo de este proyecto.

#Eliminamos columnas que no utilizaremos para liberar espacio
df_steam_games=df_steam_games.drop(columns=['url','tags','reviews_url','discount_price','metascore','specs','early_access'])

#Verificamos nuevamente las dimensiones del dataframe
df_steam_games.shape

#Tenemos solo 2 valores donde app_name no tiene valor
df_steam_games[df_steam_games['app_name'].isna()]

En lo que refiere al registro 74, podemos observar que no contamos con los datos más importantes como app_name, id, specs entre otras, intentamos a través del URL identificar el valor, pero como el URL nos dirige a un landpage y no nos permite identificar la aplicación, procedemos a eliminar el registro 74. 

df_steam_games=df_steam_games.drop(74)

En lo que refiere al registro 2580, tenemos algo más de información, como el URL, la fecha de release, id, entre otras. Vamos a proceder a buscar a partir de la URL (http://store.steampowered.com/app/317160/_/) el nombre de la app y el título, y lo vamos a actualizar en el dataframe. 

#Actualizamos tanto el app_name como el title del registro 2580
df_steam_games.loc[2580, 'app_name'] = "Duet" 
df_steam_games.loc[2580, 'title'] = "Duet"

df_steam_games.loc[2580]

Además del app_name vemos que hay 2 registros faltantes en la columna id, uno era el registro 74 eliminado, el otro es el registro 30961.

df_steam_games[df_steam_games['id'].isna()]

A partir del URL del registro 30961 (http://store.steampowered.com/app/200260) podemos determinar que el id es  200260, por lo que vamos a proceder a actualizarlo en nuestro dataframe. 

#Actualizamos el valor del id
df_steam_games.loc[30961, 'id'] = 200260

df_steam_games.loc[30961]

#Volvemos a verificar los valores nulo
df_steam_games.isnull().sum()

Podemos comprobar que tanto el id, como el release_date figuran como objeto aunque deberían ser un entero y una fecha respectivamente, procederemos a realizar dicha modificación.


#Modificamos el tipo de dato de id
df_steam_games["id"] = df_steam_games["id"].astype(int)

#Verificamos que se haya modificado el tipo de dato de la columna id
df_steam_games["id"].info()

#Modificamos el tipo de dato de release_date
df_steam_games['release_date'] = pd.to_datetime(df_steam_games['release_date'], errors='coerce')

#Verificamos que se haya modificado el tipo de dato de la columna release_date
df_steam_games["release_date"].info()

Según las funciones solicitadas, solamente vamos a necesitar del release_date, el año. Por lo tanto, crearemos una columna donde solo se incluya el año a partir de la información de la columna release_date.

#Creamos la columna release_year con el año de la columna release_date
df_steam_games["release_year"] = df_steam_games["release_date"].dt.year

#Controlamos los valores unicos que aparecen en la columna release_year
df_steam_games["release_year"].unique()

df_steam_games.shape

Luego del análisis de los datos, como varias de las funciones me solicitan que use el año  de lanzamiento, procedemos a eliminar los registros donde release_year es nulo, mantener los que no son nulos. 

#Mantenemos los valores donde el release_year sea distinto de nan.
df_steam_games = df_steam_games[df_steam_games['release_year'].notna()]

#Verificamos las dimensiones del dataframe
df_steam_games.shape

#Cambiamos el tipo de dato de la columna a entero
df_steam_games['release_year'] = df_steam_games["release_year"].astype(int)

#Verificamos el cambio de tipo de dato para la columna release_year
df_steam_games['release_year'].info()

Vamos ahora a trabajar con la columna genres.

#Verificamos la suma de los valores nulos de la columna genres
df_steam_games['genres'].isnull().sum()

En total tenemos 1234 registros que son nulos en la columna genres. Decidimos eliminarlos basándonos en que varias de las funciones nos piden un género no nulo para hacer las consultas. 

#Mantenemos los valores de genres que no son nulos
df_steam_games = df_steam_games[df_steam_games['genres'].notna()]

#Verifico las dimensiones del dataframe
df_steam_games.shape

Además, en la columna genres podemos evidenciar que tenemos una lista de elementos, por lo tanto vamos a proceder a usar el comando explode para que cada elemento de la lista se constituya en un único registro.

#Observamos como se compone la columna genres
df_steam_games['genres'].head()

#Usamos el comando explode para poder separar cada elemento de la lista en un registro
df_games = df_steam_games.explode('genres')

#Revisamos como ha quedado nuestro dataframe
df_games.head()

print('Las dimensiones del dataframe df_steam_games son:', df_steam_games.shape)
print('Las dimensiones del dataframe df_games son:', df_games.shape)

#Evidenciamos los géneros (y los vacíos)
df_games['genres'].unique()

df_games.reset_index(drop=True, inplace=True)

Finalmente decidimos eliminar otras columnas como: publisher, app_name, release_date, price y developer

#Elimino las columnas
df_games=df_games.drop(columns=['publisher','app_name', 'release_date','price','developer'])

#Verificamos el dataframe antes de exportarlo
df_games.head()

#Exportamos a csv y lo guardamos en la carpeta Data
df_games.to_csv('Data/games.csv', index=False)
