==> 02_ETL_reviews.ipynb <==
# PROYECTO INTEGRADOR: STEAM
## Dataset - australian_user_reviews.json

## 1) Ingesta de los datos

Se descomprimen cada uno de los archivos previo a importarlos con pandas.

#Importamos las librerías que utilizaremos a lo largo de nuestro proyecto
import pandas as pd
from pandas import json_normalize
from textblob import TextBlob
import json
import ast 

Importamos la data del archivo australian_user_reviews.json.

# Se carga en un dataframe las reviews de los usuarios.
user_reviews = []
with open('Data/australian_user_reviews.json', 'r', encoding='utf-8') as f:
    for line in f.readlines():
        user_reviews.append(ast.literal_eval(line))

# Crear el DataFrame anidado
df_reviews_anidado = pd.DataFrame(user_reviews)


#Observamos el dataframe
df_reviews_anidado.head()

#Imprimimos las dimensiones del dataframe
df_reviews_anidado.shape

## 2) Tratamiento de los datos

#Observamos la información contenida en un registro de la columna reviews
df_reviews_anidado['reviews'][0]

Observamos que en la columna reviews, cada registro está conformado por varios json. Por lo tanto, procederemos a desanidar la información contenida en la columna reviews.

# Desagregamos la columna reviews
df_reviews_anidado = df_reviews_anidado.explode('reviews')

#Observamos que se haya realizado la desagregación correctamente
df_reviews_anidado['reviews']

#Reseteamos los índices
df_reviews_anidado.reset_index(drop=True, inplace=True)

Vamos a trabajar con los registros nulos.

#Verificamos los nulos
df_reviews_anidado.isnull().sum()

Decido eliminar los nulos de la columna reviews, porque al ser precisamente el dataframe de reviews, es la información más importante que quiero mantener. Además, es necesario eliminarlos para un correcto funcionamiento de la API en las funciones que dependan de la columna creada en el análisis de sentimiento.

#Elimino los valores nulos de la columna reviews
df_reviews_anidado = df_reviews_anidado.dropna(subset=['reviews'])

df_reviews_anidado.isnull().sum()

#Revisamos el dataframe
df_reviews_anidado.head()

Ahora vamos a desanidar los json contenidos en la columna review.

#Aplanamos los archivos json de la columna reviews
df_reviews = json_normalize(df_reviews_anidado['reviews'])

#Revisamos como nos quedó el dataframe desanidado
df_reviews.head()

#Concateno el dataframe al que le apliqué el explode y el dataframe al que le apliqué el json_normalize
df_reviews=pd.concat([df_reviews_anidado,df_reviews], axis=1)

Vamos a revisar los valores nulos del dataframe para determinar si hay valores que se pueden eliminar.

#Revisamos los valores nulos
df_reviews.isnull().sum()

Como en las funciones, se nos solicita algún requerimiento ya sea del user_id o de las reviews (análisis de sentimiento), procederemos a  eliminar los 28 registros a partir de user_id.

#Eliminamos los valores donde user_id es nulo
df_reviews=df_reviews.dropna(subset='user_id')

#Revisamos los valores nulos nuevamente
df_reviews.isnull().sum()

Comprobamos que los 28 registros de user_id eliminados, no corresponden con los 28 registros nulos de review, por lo que ahora pasamos a eliminar los registros donde la review es nula.

#Eliminamos también valores donde la columna 'review' es vacío.
df_reviews=df_reviews.dropna(subset='review')

#Chequeamos que ya no haya valores nulos
df_reviews.isnull().sum()

#Imprimimos las dimensiones del dataframe
df_reviews.shape

#Revisamos nuestro dataframe
df_reviews.head()

Decidimos eliminar las columnas que no vamos a utilizar para las funciones o para el EDA.

#Eliminamos las columnas que no utilizaremos
df_reviews=df_reviews.drop(columns=['user_url','reviews','funny','last_edited','helpful'])

#Verificamos la correcta eliminación de las columnas
df_reviews.head()

Queremos que de la información de la columna posted podamos extraer el año (year_review) que es un dato que se nos solicita en las funciones de recomendaciones. Para ello, separamos el contenido de la columna posted en columnas y procedemos a eliminar la palabra posted y el punto final.

#Reemplazamos la palabra posted por una cadena vacía
df_reviews['posted'] = df_reviews['posted'].str.replace('Posted ', '')

#Reemplazamos el punto final por una cadena vacía
df_reviews['posted'] = df_reviews['posted'].str.replace('.', '')

#Verificamos la realización de los dos pasos anteriores.
df_reviews.head()

#Dividimos a partir de la coma la información en day_review y en year_review
df_reviews[['day_review', 'year_review']]=df_reviews['posted'].str.split(', ', expand=True)

#Revisamos la info de nuestro dataframe
df_reviews.info()

Las recomendaciones que piden las funciones, son en función de year_review, por lo tanto decidimos eliminar los valores que son nulos en esta columna.

#Tomamos los datos de df_reviews que no son nulos. 
df_reviews = df_reviews[df_reviews['year_review'].notna()]

df_reviews.info()

#Chequeamos el tipo de dato de la columna year
df_reviews['year_review'].info()

#Pasamos el dato de object a entero
df_reviews['year_review']=df_reviews['year_review'].astype(int)

#Verificamos que year_review ahora es un entero
df_reviews.info()

#Verificamos los años en los que se hicieron los posted
df_reviews['year_review'].unique()

#Verificamos las dimensiones del dataframe
df_reviews.shape

Vamos a proceder a eliminar las columnas que no utilizaremos en las funciones: posted y day_review

#Eliminamos las columnas mencionadas
df_reviews=df_reviews.drop(columns=['posted','day_review'])

#Verificamos las dimensiones del dataframe luego de eliminar algunas columnas
df_reviews.shape

#Revisamos nuestro dataframe
df_reviews.head()

#### 2.2.1) Sentiment analysis

Vamos a trabajar ahora con el dataframe df_reviews.<br>
Crearemos una columna llamada 'sentimeiento' aplicando un análisis de sentimiento con NLP tomando los siguientes valores:
- Debe tomar el valor '0' si es malo,
- Debe tomar el valor '1' si es neutral y
- Debe tomar el valor '2' si es positivo. <br>
De no ser posible este análisis por estar ausente la reseña escrita, debe tomar el valor de 1.

Para realizar el análisis de sentimiento se va a utilizar la librería textblob que es una librería de procesamiento del texto para Python que permite realizar tareas de Procesamiento del Lenguaje Natural.

Para trabajar nos vamos a centrar especificamente en la polaridad, que es un valor que va entra -1 y 1. Por lo tanto, si digo que la polaridad es negativa, el sentimiento es negativo, si es a cero es neutral (o ambiguo) y si mayor a 0, será positiva.

La consigna pide que los que estén vacíos en review se consideren para el sentiment analysis, sin embargo, como los vacíos no tienen valores en la columna recommend, se decide eliminarlos de todas maneras, para no tener nulos en las funciones que solicitan la recomendación. Además, los datos que no tienen información en year_review tampoco tienen información en review, por lo que se eliminan para simplificar las funciones en la API.

Definimos una función que al pasarle las reviews me identifique si son sentimientos positivos, neutros o negativos.

def analizar_sentimiento(texto):
    # Crear un objeto TextBlob 
    blob = TextBlob(texto)
    
    # Obtener el puntaje de sentimiento en un rango de -1 a 1
    sentimiento = blob.sentiment.polarity

    # Asignar valores numéricos según el análisis de sentimiento
    if sentimiento > 0:
        return 2  # Positivo
    elif sentimiento < 0:
        return 0  # Negativo
    else:
        return 1  # Neutral

#Aplicamos la función a la columna review del dataframe df_reviews.
df_reviews['sentimiento']=df_reviews['review'].apply(analizar_sentimiento)

#Chequeamos con los primeros 15 valores si la función tiene sentido a cómo se van clasificando los sentimientos.
df_reviews.head(15)

#Contamos para la columna sentimiento los valores
conteo_sentimiento = df_reviews['sentimiento'].value_counts()
conteo_sentimiento

#Revisamos el tipo de dato de la columna sentimiento
df_reviews['sentimiento'].info()

La columna sentimiento debe reemplazar la de user_reviews.review para facilitar el trabajo de los modelos de machine learning y el análisis de datos. Por lo tanto, procederemos a eliminar la columna review.

#Eliminamos la columna review
df_reviews=df_reviews.drop(columns=['review'])

df_reviews.head()

#Revisamos las dimensiones del dataframe
df_reviews.shape

#Guardamos en un csv
df_reviews.to_csv('Data/reviews.csv', index=False)
