==> 03_ETL_items.ipynb <==
# PROYECTO INTEGRADOR: STEAM
## Dataset -  australian_users_items.json

## 1) Ingesta de los datos

Se descomprimen cada uno de los archivos previo a importarlos con pandas.

#Importamos las librerías que utilizaremos a lo largo de nuestro proyecto
import pandas as pd
from pandas import json_normalize
import json
import ast 

Importamos la información del archivo australian_users_items.json.

# Se carga en un dataframe los items de los usuarios.
users_items = []
with open('Data/australian_users_items.json', 'r', encoding='utf-8') as f:
    for line in f.readlines():
        users_items.append(ast.literal_eval(line))

# Crear el DataFrame anidado
df_users_items = pd.DataFrame(users_items)


#Observamos el dataframe
df_users_items.head()

#Imprimimos las dimensiones del dataframe
df_users_items.shape

## 2) Tratamiento de los datos

#Revisamos el dataframe
df_users_items.head()

#Revisamos la info del dataframe
df_users_items.info()

#Elimino las columnas que no voy a utilizar
df_users_items=df_users_items.drop(columns=['steam_id', 'user_url','items_count'], axis=1)

Queremos desanidar la columna items que está conformada por json anidados.Para eso primero aplicamos el comando explode para expandir o explotar las columnas

# Desagregamos la columna items
df_items_anidado = df_users_items.explode('items')


#Reseteo el index
df_items_anidado.reset_index(drop=True, inplace=True)

#Reviso que se haya realizado la modificación
df_items_anidado.head(5)

#Observo  si hay valores nulos
df_items_anidado.isnull().sum()

Vamos a proceder a hacer el json_normalize para desanidar items, previo a tratar los valores nulos.

#Aplico la normalización a la columna items
df_items_norm = json_normalize(df_items_anidado['items'])

#Reviso que se haya aplanado la columna items
df_items_norm

#Elimino la columna playtime_2weeks que no la voy a utilizar
df_items_norm=df_items_norm.drop(columns=['playtime_2weeks'])

#Verifico las dimensiones del dataframe
df_items_norm.shape

#Concateno el dataframe al que le hice el explode con el dataframe al que le apliqué la normalización
df_items=pd.concat([df_items_anidado,df_items_norm], axis=1)

#Verificamos las dimensiones de nuestro dataframe
df_items.shape

#Reviso los nulos nuevamente
df_items.isnull().sum()

Procedemos a eliminar los nulos. Precisamente en el dataframe items, me interesa tener información sobre los items por lo los registros que solo tienen usuarios, los elimino sino no podré hacer correctamente los merge necesarios para las funciones. 

#Eliminamos los nulos de la columna items
df_items=df_items.dropna(subset='items')

#Verificamos que ya no haya nulos
df_items.isnull().sum()

#Revisamos las dimensiones del dataframe
df_items.shape

#Eliminamos la columna items (anidada)
df_items=df_items.drop(columns=['items'])

#Verificamos la eliminación
df_items.head()

Para las funciones solicitadas, se solicita el tiempo de juego por usuario o por género. En este paso y en función de la cantidad de datos presentes, se decide eliminar los valores donde playtime_forever es 0 considerando unicamente las reviews de los usuarios que jugaron al menos 1 hora. El diccionario de datos no determina ninguna cuestión respecto de los usuarios que juegan el juego por menos de una hora, por lo que asumiremos que esos casos son considerados como 0 y también se eliminan.

#Me quedo con los valores donde playtime_forever es distinto de cero
df_items=df_items[df_items['playtime_forever']!=0]

#Reseteamos los indices
df_items.reset_index(drop=True, inplace=True)

#Verificamos nuevamente las dimensiones del dataframe
df_items.shape

#Verificamos el tipo de dato de la columna playtime
df_items['playtime_forever'].info()

#Pasamos de flotante a entero
df_items['playtime_forever'] = df_items['playtime_forever'].astype(int)

#Verificamos la modificación
df_items['playtime_forever'].info()

#Exportamos a csv
df_items.to_csv('Data/items.csv',index=False)
