==> 04_Tratamiento_para_funciones.ipynb <==
# PROYECTO INTEGRADOR: STEAM
## Tratamiento de los datasets para las funciones

En este notebook se presentan las modificaciones y relaciones entre los datasets obtenidos por el ETL de manera que nos permiten optimizar las funciones requeridas. 


Se intentará en cada caso (y en la medida de lo posible) filtrar las columnas y los registros requeridos para las funciones en este archivo, de manera de que la información llegue lo más limpia posible a la API y no haya inconvenientes por la cantidad de datos.

Para cada función en específico, plantearemos un dataset lo más óptimo posible, que exportaremos desde este archivo pero que importaremos luego en el main.py necesario para fastAPI.

#Importamos librerías
import pandas as pd
import ast

#Importamos el dataset de games
df_games=pd.read_csv('Data/games.csv')
df_games.head()

#Importamos el dataset de reviews
df_reviews=pd.read_csv('Data/reviews.csv')
df_reviews.head()

#Importamos el dataset de items
df_items=pd.read_csv('Data/items.csv')
df_items.head()

## **1) Función PlayGenRe**

def PlayTimeGenre( genero : str ): Debe devolver año con mas horas jugadas para dicho género.
#Ejemplo de retorno: {"Año de lanzamiento con más horas jugadas para Género X" : 2013}

En este caso, vamos a considerar un dataset que solo contengan los géneros, el release_year (año de lanzamiento) y las horas jugadas (playtime_forever). Lo llamaremos: dataset_generos_horas_jugadas.

Como en el dataframe de games tenemos la información correspondiente a los géneros y a los años de lanzamiento, y en el dataframe de items poseemos la información de las horas de juego, vamos a hacer un merge entre ambos dataframe basándonos en el item_id (que el df_games se llama id, y en df_items se llama item_id)

#Se hace un merge entre los dos dataframes
df_generos_horas_jugadas = pd.merge(df_games, df_items, left_on='id', right_on='item_id', how='inner')


#Observamos como nos quedan los datos
df_generos_horas_jugadas.head()

#Eliminamos las columnas que no vamos a utilizar, y nos quedamos solo con las mencionadas anteriormente.
df_generos_horas_jugadas=df_generos_horas_jugadas.drop(columns=['title','id','item_id','item_name','user_id'])

df_generos_horas_jugadas.head()

Como una manera de simplificar la función, modificamos nuestro dataset de manera que funcione de una manera más óptima en el render. Lo que se procede a hacer es, sabiendo que para cada género dado queremos que la función nos sume el playtime_forever por año, por lo que haremos un group by.

#Se agrupa en función de genres y release_year, sumando los playtime_forever y reseteando los índices 
df_generos_horas_jugadas = df_generos_horas_jugadas.groupby(['genres','release_year'])['playtime_forever'].sum().reset_index()

#Verificamos que se hayan agrupado correctamente
df_generos_horas_jugadas

Logramos disminuir nuestro dataset a 3 columnas y 340 filas.

Este dataset lo usaremos para la primera función, por lo que procederemos a exportarlo en csv, para utilizarlo en nuestro archivo main.py

#Exportamos a csv
df_generos_horas_jugadas.to_csv('Data/generos_horas_jugadas.csv', index=False)

## **2) Función UserFonGenre**

def UserForGenre( genero : str ): Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año.
#Ejemplo de retorno: {"Usuario con más horas jugadas para Género X" : us213ndjss09sdf, "Horas jugadas":[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas: 23}]}


Para crear el dataframe para esta función haremos un procedimiento parecido al realizado anteriormente. 
En este caso, nos piden que al ingresar un género, la función nos dé el usuario con más horas acumuladas, seguido de una lista de acumulación de horas jugadas por año. 

Los datos son parecidos al caso anterior, necesitamos tomar del df_games los valores de genres y release_year, y del dataframe de items, los valores de playtime_forever y de user_id, por eso procedemos a hacer un merge entre ambos dataframes

#Hacemos un merge entre df_games y df_items
df_usuario_horas_jugadas = pd.merge(df_games, df_items, left_on='id', right_on='item_id', how='inner')

#Revisamos los datos
df_usuario_horas_jugadas.head()

#Eliminamos las columnas que no vamos a utilizar.
df_usuario_horas_jugadas.drop(columns=['title','id','item_id','item_name'])

Como en la primer función, en este caso vamos a aprovechar a realizar el agrupamiento para que se sumen las horas acumuladas y así disminuir el tamaño de nuestro dataset. 

#Hacemos group by sumando los playtime
df_usuario_horas_jugadas = df_usuario_horas_jugadas.groupby(['genres', 'release_year','user_id'])['playtime_forever'].sum().reset_index()

#Revisamos el group by
df_usuario_horas_jugadas

#Exportamos a csv
#df_usuario_horas_jugadas.to_csv('Data/usuario_horas_jugadas.csv', index=False)

Al exportar a csv, el archivo pesa mas que los 50mb recomendados por archivos para github. Por lo tanto decidimos pasarlo a parquet.

df_usuario_horas_jugadas.to_parquet('Data/usuario_horas_jugadas.parquet', engine='pyarrow')

## **3) Función UsersRecommend**

def UsersRecommend( año : int ): Devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = True y comentarios positivos/neutrales)
#Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

Para este caso, queremos que al ingresar un año en especial, basado en que las recomendaciones sean True, y los review sean positivas o neutrales, nos recomienda los top 3 juegos más recomendados. 

Lo que deberíamos hacer es hacer un merge entre df_reviews de donde tomaremos la columna year_review, recommend y sentimiento (columna creada a partir del analisis de sentimiento) y debemos hacer un merge para obtener el nombre de los titulos de los juegos.

Como una manera de optimizar la consulta, como podría hacer un merge de df_reviews con df_games o con df_items, porque en ambas tengo los títulos, lo haré con df_games que tiene menos datos para optimizar el tiempo de ejecución.

#Hacemos el merge entre df_games y df_reviews en función de la columna id e item_id que tienen la misma información
df_recomendaciones_positivas=pd.merge(df_games, df_reviews, left_on='id', right_on='item_id', how='inner')

#Observamos como nos queda el df
df_recomendaciones_positivas.head()

Elijo quedarme con la columna de year_review, en vez de release_year, porque como vamos a considerar las recomendaciones, el año de la recomendación es mucho más acertado para el análisis que deseamos hacer (puede que en el año de salida del juego aún no haya recomendaciones). Procedemos a eliminar release_year y otras columnas que no utilizaremos para esta función.

#Eliminamos las columnas que no utilizaremos para esta función
df_recomendaciones_positivas=df_recomendaciones_positivas.drop(columns=['genres','id', 'release_year','user_id','item_id'])

df_recomendaciones_positivas.head()

#Contamos los valores True y False de la columna recommend
df_recomendaciones_positivas['recommend'].value_counts()

Para optimizar aún más la consulta, vamos a filtrar nuestro dataframe y vamos a quedarnos en primera instancia con los recommend=True para las recomendaciones positivas.

#Para este caso, me elijo con los recommend que sean True, y los filtro
df_recomendaciones_positivas=df_recomendaciones_positivas[df_recomendaciones_positivas['recommend']==True]

#Vuelvo a hacer el conteo, para verificar que mi df solo tenga los valores True
df_recomendaciones_positivas['recommend'].value_counts()

Ahora también vamos a filtrar los valores donde sentimiento sea 1 o 2 (neutral o positivo) por lo que eliminamos donde la columna sentimiento sea 0.

#Verifico ahora, para los recommend = True, el conteo de los valores de la columna sentimiento
df_recomendaciones_positivas['sentimiento'].value_counts()

#Me quedo solo en mi df con los valores de sentimiento que sean distintos de 0, 1 y 2 (positivo y neutral)
df_recomendaciones_positivas=df_recomendaciones_positivas[df_recomendaciones_positivas['sentimiento']!=0]

#Verifico el conteo de los valores nuevamente para ver el cambio
df_recomendaciones_positivas['sentimiento'].value_counts()

#Revisamos el df
df_recomendaciones_positivas

Hemos optimizado de la mejor manera posible el dataframe, ahora lo exportaremos para usar en las funciones

#Exportamos a csv
df_recomendaciones_positivas.to_csv('Data/recomendaciones_positivas.csv', index=False)

 ## **4) Función UsersNotRecommend**

def UsersNotRecommend( año : int ): Devuelve el top 3 de juegos MENOS recomendados por usuarios para el año dado. (reviews.recommend = False y comentarios negativos)
#Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

En este caso, es análogo al anterior, pero al hacer el merge entre df_reviews y df_games, me voy a quedar con las recomendaciones que sean falsas, y los sentimiento igual a 0. 

#Hago el merge entre df_games y df_reviews a través de la columna id e item_id que tienen la misma información
df_recomendaciones_negativas=pd.merge(df_games, df_reviews, left_on='id', right_on='item_id', how='inner')

#Vemos el resultado del merge
df_recomendaciones_negativas

Como en el caso anterior, para basarme en un análisis de recomendaciones, considero el year_review y no el year_release por lo que procederé a eliminarlo.

#Eliminamos las columnas que no utilizaremos para esta función
df_recomendaciones_negativas=df_recomendaciones_negativas.drop(columns=['genres','id', 'release_year','user_id','item_id'])

#Cuento los valores que tengo en la columna recommend
df_recomendaciones_negativas['recommend'].value_counts()

#El df de recomendaciones negativas, solo mantendrá los valores donde recommend sea falso
df_recomendaciones_negativas=df_recomendaciones_negativas[df_recomendaciones_negativas['recommend']==False]

#Verifico que me haya quedado con los valores falsos
df_recomendaciones_negativas['recommend'].value_counts()

Ahora vamos a optimizar la columna de sentimiento, para un dataframe que solo contiene las recomendaciones que son False y vamos a intentar quedarnos con los valores donde sentimiento = 0.

#Ahora reviso que para recommend=False, los valores de la columna sentimiento
df_recomendaciones_negativas['sentimiento'].value_counts()

#Para este df, solo mantengo los valores de la columna sentimiento que sean 0
df_recomendaciones_negativas=df_recomendaciones_negativas[df_recomendaciones_negativas['sentimiento']==0]

#Verifico que se hayan eliminado los valores
df_recomendaciones_negativas['sentimiento'].value_counts()

#Reviso como me queda el dataframe
df_recomendaciones_negativas

#Exportamos a csv
df_recomendaciones_negativas.to_csv('Data/recomendaciones_negativas.csv', index=False)

## **5) Función sentiment_analysis**

def sentiment_analysis( año : int ): Según el año de lanzamiento, se devuelve una lista con la cantidad de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento.
#Ejemplo de retorno: {Negative = 182, Neutral = 120, Positive = 278}

Para este caso, necesito que al ingresar el año, me devuelva una lista con la cantidad de registros según sentimiento que haya en cada caso. Por lo ranto, necesito que mi dataframe específico para esta función tenga el nombre del juego, el year_review y el sentimiento.

En mi df_reviews, tomaré el year_review y el sentimiento, pero me faltan los títulos, así que haré un merge, en este caso con df_games, que es un dataset más pequeño y me permite optimizar la operación.

#Hago el merge entre df_games y df_reviews a partir de la columna id e item_id que tienen la misma información
df_sentiment_analysis=pd.merge(df_games, df_reviews, left_on='id', right_on='item_id', how='inner')

#Verifico el dataset obtenido
df_sentiment_analysis.head()

Usaremos el año de lanzamiento (release_year) porque el enunciado lo pide. Particularmente creo que tendría más sentido hacerlo con el year_review

#Eliminamos las columnas que no usaremos
df_sentiment_analysis=df_sentiment_analysis.drop(columns=['genres','id', 'year_review','user_id','item_id','recommend','title'])

#Verificamos que las columnas se hayan elimnado correctamente
df_sentiment_analysis

#Exportamos a csv
df_sentiment_analysis.to_csv('Data/sentiment_analysis.csv', index=False)
